requirements.txt
mock
nose
pycurl
py-bcrypt
pika
redis
PyYAML
protobuf
flask
python-consul
prometheus_client
requests
